fix componentdao

+++ src/persistence/ComponentOracleDaoImpl.java
@@ -10,7 +10,7 @@ import domain.Wagon;
 
 public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDao {
 	public boolean createLocomotive(Locomotive locomotive) {
-		try(Connection conn = super.getConnection()) {
+		try (Connection conn = super.getConnection()) {
 			String query = "insert into components (name, type) values(?, ?)";
 			PreparedStatement pstmt = conn.prepareStatement(query);
 			pstmt.setString(1, locomotive.getName());
@@ -22,9 +22,9 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return false;
 		}
 	}
-	
+
 	public boolean createWagon(Wagon wagon) {
-		try(Connection conn = super.getConnection()) {
+		try (Connection conn = super.getConnection()) {
 			String query = "insert into components (name, seats, type) values(?, ?, ?)";
 			PreparedStatement pstmt = conn.prepareStatement(query);
 			pstmt.setString(1, wagon.getName());
@@ -37,10 +37,10 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return false;
 		}
 	}
-	
+
 	public boolean addComponent(String trainName, String componentName) {
 		int place = checkComponentPlace(trainName);
-		try(Connection conn = super.getConnection()) {
+		try (Connection conn = super.getConnection()) {
 			String query = "insert into connection (trainname, componentname, place) values(?, ?, ?)";
 			PreparedStatement pstmt = conn.prepareStatement(query);
 			pstmt.setString(1, trainName);
@@ -53,25 +53,29 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return false;
 		}
 	}
-	
-	//teller van plaats connection aanpassen
+
 	public boolean removeComponent(String trainName, String componentName) {
 		int place = getLastComponentInConnection(trainName, componentName);
-		try(Connection conn = super.getConnection()) {
-			String query = "delete from connection where trainname = ? and componentname = ? and place = ?";
-			PreparedStatement pstmt = conn.prepareStatement(query);
-			pstmt.setString(1, trainName);
-			pstmt.setString(2, componentName);
-			pstmt.setInt(3, place);
-			pstmt.executeQuery();
-			lowerPlaceComponents(place, trainName);
-			return true;
-		} catch (SQLException e) {
-			e.printStackTrace();
+		if (place == -1) {
+			System.out.println("Kan component niet verwijderen");
 			return false;
+		} else {
+			try (Connection conn = super.getConnection()) {
+				String query = "delete from connection where trainname = ? and componentname = ? and place = ?";
+				PreparedStatement pstmt = conn.prepareStatement(query);
+				pstmt.setString(1, trainName);
+				pstmt.setString(2, componentName);
+				pstmt.setInt(3, place);
+				pstmt.executeQuery();
+				lowerPlaceComponents(place, trainName);
+				return true;
+			} catch (SQLException e) {
+				e.printStackTrace();
+				return false;
+			}
 		}
 	}
-	
+
 	public boolean deleteComponent(String componentName) {
 		if (checkComponentUsed(componentName)) {
 			try (Connection conn = super.getConnection()) {
@@ -89,7 +93,7 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return false;
 		}
 	}
-	
+
 	public int getSeatsComponent(String name) {
 		try (Connection conn = super.getConnection()) {
 			String query = "select seats from component where name = ?";
@@ -104,7 +108,7 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return 0;
 		}
 	}
-	
+
 	private int checkComponentPlace(String trainName) {
 		try (Connection conn = super.getConnection()) {
 			String sql = "select count(trainname) amountcomponents from connection where trainname = ?";
@@ -119,7 +123,7 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			return -1;
 		}
 	}
-	
+
 	private boolean checkComponentUsed(String componentName) {
 		try (Connection conn = super.getConnection()) {
 			String sql = "select componentname from connection where componentname = ?";
@@ -133,14 +137,14 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			} else {
 				System.out.println("Component bestaat niet in connection");
 				return true;
-			} 
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 			System.out.println("sql error checkComponentUsed()");
 			return false;
 		}
 	}
-	
+
 	private int getLastComponentInConnection(String trainName, String componentName) {
 		try (Connection conn = super.getConnection()) {
 			String sql = "select place from connection where trainname = ? and componentname = ? order by place desc fetch first 1 rows only";
@@ -149,13 +153,18 @@ public class ComponentOracleDaoImpl extends OracleBaseDao implements ComponentDa
 			pstmt.setString(2, componentName);
 			ResultSet rs = pstmt.executeQuery();
 			rs.next();
-			return rs.getInt("place");
+			if (rs.next()) {
+				return rs.getInt("place");
+			} else {
+				System.out.println("Component bestaat niet!");
+				return -1;
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 			return -1;
 		}
 	}
-	
+
 	private void lowerPlaceComponents(int place, String trainName) {
 		try (Connection conn = super.getConnection()) {
 			String sql = "update connection set place = place - 1 where place > ? and trainname = ?";
